[SYSTEMDS-3465] Typed return on CacheBlock Interface Slice
This commit change the CacheBlock interfaced to by Typed.
The change allows cleaner interaction of instructions that are applied to
Tensors, Matrices, and Frames such as slice. where previously any
cacheBlock could be the argument target for the slice command and
potentially any other type could be returned. Now it only allows equal
type input and return equal type out.

Also cleaned up in this commit is the memory estimates for FrameBlocks,
to continue the process of separating the responsibility of the
ArrayColumns from the FrameBlock.

There are also performance benefits in this commit slightly improving
the parsing of Pandas DataFrame into SystemDS DataFrames, but further
improvements are planed in future commits.

+++ src/main/java/org/apache/sysds/runtime/compress/colgroup/dictionary/QDictionary.java
@@ -93,7 +93,7 @@ public class QDictionary extends ADictionary {
 
 	public static long getInMemorySize(int valuesCount) {
 		// object + values array + double
-		return 16 + MemoryEstimates.byteArrayCost(valuesCount) + 8;
+		return 16 + (long)MemoryEstimates.byteArrayCost(valuesCount) + 8;
 	}
 
 	@Override
